feat(icons): add icons in multiple places
new icons for: Write, Preview, Post, Github link, New reaction and some basic tidy up of the
reaction icons
