+++ src/app/editor/editor.component.html
@@ -1,7 +1,17 @@
 <div class="flex flex-col border rounded-lg shadow-lg overflow-auto mb-8 bg-white">
   <div class="flex flex-row px-4 pt-2 border-b">
-    <button (click)="write()" [ngClass]="getTabCss(!previewing)"  class="bg-white py-2 px-4 font-semibold inline-block">Write</button>
-    <button (click)="preview()" [ngClass]="getTabCss(previewing)" class="bg-white py-2 px-4 font-semibold inline-block">Preview</button>
+    <button (click)="write()" [ngClass]="getTabCss(!previewing)"  class="bg-white py-2 px-4 font-semibold inline-block">
+      <span class="flex justify-center items-center gap-1">
+        <span [innerHTML]="writeIcon | safe" class="w-6 h-6 inline-block"></span>
+        Write
+      </span>
+    </button>
+    <button (click)="preview()" [ngClass]="getTabCss(previewing)" class="bg-white py-2 px-4 font-semibold inline-block">
+      <span class="flex justify-center items-center gap-1">
+        <span [innerHTML]="previewIcon | safe" class="w-6 h-6 inline-block"></span>
+        Preview
+      </span>
+    </button>
   </div>
   <textarea
     class="border-none text-gray-500 min-h-[10rem] focus:border-none p-2"
@@ -17,7 +27,10 @@
       [disabled]="!comment"
       (click)="post()"
     >
-      Post
+      <span class="flex justify-center items-center gap-1">
+        <span [innerHTML]="postIcon | safe" class="w-6 h-6 inline-block"></span>
+        Post
+      </span>
     </button>
   </div>
 </div>
