+++ src/app/new-reaction/new-reaction.component.html
@@ -5,7 +5,8 @@
       (addReaction)="addReaction.emit($event)"
     ></app-reaction>
   </div>
-  <button class="rounded-full border px-2 bg-white hover:bg-slate-200 flex flex-row justify-center items-center gap-1">
-    <span class="text-slate-500">+</span>
+  <button class="text-blue-500 rounded-full border border-blue-500 px-2 bg-white hover:bg-slate-200 flex flex-row justify-center items-center gap-1 h-6"
+    title="Add reaction">
+    <span class="w-5 h-5" [innerHTML]="chevronIcon | safe"></span>
   </button>
 </span>
