+++ src/app/new-reaction/new-reaction.component.spec.ts
@@ -1,4 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { SafePipe } from '../safe.pipe';
 
 import { NewReactionComponent } from './new-reaction.component';
 
@@ -8,7 +9,7 @@ describe('NewReactionComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ NewReactionComponent ]
+      declarations: [ NewReactionComponent, SafePipe ]
     })
     .compileComponents();
