+++ src/app/reaction/reaction.component.html
@@ -1,5 +1,7 @@
-<button class="rounded-full border px-2 bg-white hover:bg-slate-200 flex flex-row justify-center items-center gap-1"
+<button 
+  class="rounded-full border px-2 bg-white hover:bg-slate-200 
+    flex flex-row justify-center items-center gap-1 h-6 text-slate-500"
   (click)="addReaction.emit(reaction)">
-  <span class="w-5 h-5" [innerHtml]="reactionSVG | safe"></span>
-  <span *ngIf="count" class="text-slate-500">{{ count }}</span>
+  <span [innerHtml]="reactionSVG | safe" class="w-5 h-5"></span>
+  <span *ngIf="count">{{ count }}</span>
 </button>
